Upstream version 1.0.0~0alpha

+++ src/test/compile-fail/bad-env-capture.rs
@@ -8,9 +8,9 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-// error-pattern: attempted dynamic environment-capture
+// error-pattern: can't capture dynamic environment in a fn item;
 fn foo() {
-    let x: int;
+    let x: isize;
     fn bar() { log(debug, x); }
 }
 fn main() { foo(); }
