+++ src/test/compile-fail/borrowck-anon-fields-struct.rs
@@ -1,7 +1,17 @@
-// Tests that we are able to distinguish when loans borrow different
-// anonymous fields of an tuple vs the same anonymous field.
+// Copyright 2014 The Rust Project Developers. See the COPYRIGHT
+// file at the top-level directory of this distribution and at
+// http://rust-lang.org/COPYRIGHT.
+//
+// Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
+// http://www.apache.org/licenses/LICENSE-2.0> or the MIT license
+// <LICENSE-MIT or http://opensource.org/licenses/MIT>, at your
+// option. This file may not be copied, modified, or distributed
+// except according to those terms.
 
-struct Y(uint, uint);
+// Tests that we are able to distinguish when loans borrow different
+// anonymous fields of a tuple vs the same anonymous field.
+
+struct Y(usize, usize);
 
 fn distinct_variant() {
     let mut y = Y(1, 2);
