+++ src/test/compile-fail/capture1.rs
@@ -1,4 +1,3 @@
-// -*- rust -*-
 // Copyright 2012 The Rust Project Developers. See the COPYRIGHT
 // file at the top-level directory of this distribution and at
 // http://rust-lang.org/COPYRIGHT.
@@ -10,9 +9,9 @@
 // except according to those terms.
 
 
-// error-pattern: attempted dynamic environment-capture
+// error-pattern: can't capture dynamic environment in a fn item;
 
 fn main() {
-    let bar: int = 5;
-    fn foo() -> int { return bar; }
+    let bar: isize = 5;
+    fn foo() -> isize { return bar; }
 }
