+++ src/test/compile-fail/issue-2478.rs
@@ -1,4 +1,4 @@
-// Copyright 2012 The Rust Project Developers. See the COPYRIGHT
+// Copyright 2012-2014 The Rust Project Developers. See the COPYRIGHT
 // file at the top-level directory of this distribution and at
 // http://rust-lang.org/COPYRIGHT.
 //
@@ -8,9 +8,10 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-// xfail-test
-fn foo<'a>() -> &'a int {  //~ ERROR unconstrained region
+// ignore-test
+
+fn foo<'a>() -> &'a isize {  //~ ERROR unconstrained region
     return &x;
 }
-static x: int = 5;
+static x: isize = 5;
 fn main() {}
