+++ src/test/compile-fail/not-a-pred.rs
@@ -1,4 +1,3 @@
-// -*- rust -*-
 // Copyright 2012 The Rust Project Developers. See the COPYRIGHT
 // file at the top-level directory of this distribution and at
 // http://rust-lang.org/COPYRIGHT.
@@ -11,8 +10,8 @@
 
 // error-pattern: lt
 
-fn f(a: int, b: int) : lt(a, b) { }
+fn f(a: isize, b: isize) : lt(a, b) { }
 
-fn lt(a: int, b: int) { }
+fn lt(a: isize, b: isize) { }
 
-fn main() { let a: int = 10; let b: int = 23; check (lt(a, b)); f(a, b); }
+fn main() { let a: isize = 10; let b: isize = 23; check (lt(a, b)); f(a, b); }
