+++ src/test/compile-fail/type-shadow.rs
@@ -1,4 +1,3 @@
-// -*- rust -*-
 // Copyright 2012 The Rust Project Developers. See the COPYRIGHT
 // file at the top-level directory of this distribution and at
 // http://rust-lang.org/COPYRIGHT.
@@ -10,7 +9,7 @@
 // except according to those terms.
 
 fn main() {
-    type X = int;
+    type X = isize;
     type Y = X;
     if true {
         type X = &'static str;
